move discount deletes only to the discount service
